+++ crates/edr_napi/src/mock/time.rs
@@ -100,7 +100,9 @@ pub fn create_provider_with_mock_timer<'env>(
             )?;
 
             let provider_config =
-                edr_provider::ProviderConfig::<edr_chain_l1::Hardfork>::try_from(provider_config)?;
+                edr_provider::config::Provider::<edr_chain_l1::Hardfork>::try_from(
+                    provider_config,
+                )?;
 
             let provider =
                 edr_provider::Provider::<GenericChainSpec, Arc<edr_provider::time::MockTime>>::new(