+++ crates/edr_op/tests/integration/provider.rs
@@ -7,9 +7,10 @@ use edr_eth::BlockSpec;
 use edr_op::OpChainSpec;
 use edr_primitives::{address, bytes, Address, HashMap, U64};
 use edr_provider::{
+    config::{ForkConfig, ProviderConfig},
     test_utils::{create_test_config_with, MinimalProviderConfig, ProviderTestFixture},
     time::CurrentTime,
-    ForkConfig, MethodInvocation, NoopLogger, Provider, ProviderConfig, ProviderRequest,
+    MethodInvocation, NoopLogger, Provider, ProviderRequest,
 };
 use edr_solidity::contract_decoder::ContractDecoder;
 use edr_test_utils::env::json_rpc_url_provider;